# Conflicts:
#	Assets/02.Prefabs/Monster/MonsterType1.prefab
#	Assets/02.Prefabs/Monster/MonsterType2.prefab
#	Assets/02.Prefabs/Monster/MonsterType3.prefab

+++ Assets/01.Scripts/DataTable/MonsterTable.cs
@@ -19,7 +19,7 @@ public class MonsterData
     public int Monster_Gold { get; set; }
     public int Monster_Skill_Id { get; set; }
     public float Cooldown { get; set; }
-    public int Range { get; set; }
+    public float Range { get; set; }
 }
 
 public class MonsterTable : DataTable
